+++ web/src/lib/components/timeline/Timeline.svelte
@@ -250,7 +250,7 @@
       scrollToSegmentPercentage(0, timelineManager.topSectionHeight, scrubberMonthScrollPercent);
     } else if (leadOut) {
       scrollToSegmentPercentage(
-        timelineManager.topSectionHeight + timelineManager.assetsHeight,
+        timelineManager.topSectionHeight + timelineManager.bodySectionHeight,
         timelineManager.bottomSectionHeight,
         scrubberMonthScrollPercent,
       );
@@ -611,7 +611,7 @@
       style:position="absolute"
       style:left="0"
       style:right="0"
-      style:transform={`translate3d(0,${timelineManager.topSectionHeight + timelineManager.assetsHeight}px,0)`}
+      style:transform={`translate3d(0,${timelineManager.topSectionHeight + timelineManager.bodySectionHeight}px,0)`}
     ></div>
   </section>
 </section>
